create custom dataclass decorator
