+++ simple_pytree/__init__.py
@@ -1,5 +1,6 @@
 __version__ = "0.1.7"
 
-from .pytree import Pytree, field, static_field
+from .dataclass import dataclass, field, static_field
+from .pytree import Pytree, PytreeMeta
 
-__all__ = ["Pytree", "field", "static_field"]
+__all__ = ["Pytree", "PytreeMeta", "dataclass", "field", "static_field"]
