+++ simple_pytree/dataclass.py
@@ -1,0 +1,103 @@
+import dataclasses
+import typing as tp
+
+import typing_extensions as tpe
+
+A = tp.TypeVar("A")
+
+
+def field(
+    *,
+    default: tp.Any = dataclasses.MISSING,
+    pytree_node: bool = True,
+    default_factory: tp.Any = dataclasses.MISSING,
+    init: bool = True,
+    repr: bool = True,
+    hash: tp.Optional[bool] = None,
+    compare: bool = True,
+    metadata: tp.Optional[tp.Mapping[str, tp.Any]] = None,
+):
+    if metadata is None:
+        metadata = {}
+    else:
+        metadata = dict(metadata)
+
+    if "pytree_node" in metadata:
+        raise ValueError("'pytree_node' found in metadata")
+
+    metadata["pytree_node"] = pytree_node
+
+    return dataclasses.field(  # type: ignore
+        default=default,
+        default_factory=default_factory,
+        init=init,
+        repr=repr,
+        hash=hash,
+        compare=compare,
+        metadata=metadata,
+    )
+
+
+def static_field(
+    *,
+    default: tp.Any = dataclasses.MISSING,
+    default_factory: tp.Any = dataclasses.MISSING,
+    init: bool = True,
+    repr: bool = True,
+    hash: tp.Optional[bool] = None,
+    compare: bool = True,
+    metadata: tp.Optional[tp.Mapping[str, tp.Any]] = None,
+):
+    return field(
+        default=default,
+        pytree_node=False,
+        default_factory=default_factory,
+        init=init,
+        repr=repr,
+        hash=hash,
+        compare=compare,
+        metadata=metadata,
+    )
+
+
+@tp.overload
+def dataclass(cls: tp.Type[A]) -> tp.Type[A]:
+    ...
+
+
+@tp.overload
+def dataclass(
+    *,
+    init: bool = True,
+    repr: bool = True,
+    eq: bool = True,
+    order: bool = False,
+    unsafe_hash: bool = False,
+    frozen: bool = False,
+) -> tp.Callable[[tp.Type[A]], tp.Type[A]]:
+    ...
+
+
+@tpe.dataclass_transform(field_specifiers=(field, static_field, dataclasses.field))
+def dataclass(
+    cls: tp.Optional[tp.Type[A]] = None,
+    init: bool = True,
+    repr: bool = True,
+    eq: bool = True,
+    order: bool = False,
+    unsafe_hash: bool = False,
+    frozen: bool = False,
+) -> tp.Union[tp.Type[A], tp.Callable[[tp.Type[A]], tp.Type[A]]]:
+    decorator = dataclasses.dataclass(
+        init=init,
+        repr=repr,
+        eq=eq,
+        order=order,
+        unsafe_hash=unsafe_hash,
+        frozen=frozen,
+    )
+
+    if cls is None:
+        return decorator
+
+    return decorator(cls)
